% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{taxids_available_lists}
\alias{taxids_available_lists}
\title{Describe available datasets}
\usage{
taxids_available_lists()
}
\value{
vector of available datasets
}
\description{
Describe available datasets
}
\examples{
taxids_available_lists()
}
