% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{taxids_load_list}
\alias{taxids_load_list}
\title{Load a dataset}
\usage{
taxids_load_list(taxid_list_to_load = NULL, verbose = FALSE)
}
\arguments{
\item{taxid_list_to_load}{a taxid list to load. See available_taxid_files_no_extension for options. (string)}

\item{verbose}{verbose (boolean)}
}
\value{
vector of taxids
}
\description{
Load a dataset
}
\examples{
\dontrun{
taxids_load_list()
}
}
